Move implementation of fcntl to OSUtil dir & resolve comments
